Refactor: búsqueda se carga inline en la operación de compra, no en el lead
- Elimina tipoLead y la relación @OneToOne busqueda de Lead
- Elimina TipoLead enum
- Elimina endpoint PUT /leads/{id}/busqueda y su método de servicio
- asociarBusquedaACompra ahora crea la búsqueda en el mismo acto que la operación
- Corrige tipo genérico de OperacionRepository de Integer a Long

+++ src/main/java/com/leadera/leadera/repository/OperacionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface OperacionRepository extends JpaRepository<Operacion, Integer> {
+public interface OperacionRepository extends JpaRepository<Operacion, Long> {
     List<Operacion> findByLeadIdAndAgenteEmail(Long leadId, String email);
 
     List<Operacion> findByLeadIdAndAgenteEmailAndEstadoOperacionNotIn(
